backend/requirements.txt
Flask
flask-cors
Pillow>=11
numpy>=2.1